+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.4] - 2026-08-06
+
+### Added
+
+- search-debian and search-ubuntu to search for ABIs across codenames.
+- Output JSON for analysis.
+
 ## [1.1.3] - 2026-08-06
 
 ### Added